14 Aplicação Avançadas com Angular/projetos construidos/jv-state-ngrx-master/src/app/features/dashboard/list-item/list-item.component.ts
import { Component, ChangeDetectionStrategy, Input, Output, EventEmitter } from '@angular/core';

import { Todo } from 'src/app/shared/models/todo.model';

@Component({
  selector: 'jv-list-item',
  templateUrl: './list-item.component.html',
  styleUrls: ['./list-item.component.scss'],
  changeDetection: ChangeDetectionStrategy.OnPush,
})
export class ListItemComponent {

  @Input() item: Todo;
  @Output() markAsDone = new EventEmitter<number>();
  @Output() remove = new EventEmitter<number>();
}
